data_pipeline/dags/requirements.txt
pandas
scikit-learn
kneed
google-cloud-bigquery
apache-airflow-providers-google
jinja2